refactor(repository): refactored v1 encryption overhead to be a function that's only invoked when actual V1 index is opened

+++ repo/content/content_manager_indexes.go
@@ -88,7 +88,7 @@ func ParseIndexBlob(ctx context.Context, blobID blob.ID, encrypted gather.Bytes,
 		return nil, errors.Wrap(err, "unable to decrypt index blob")
 	}
 
-	ndx, err := index.Open(data.Bytes().ToByteSlice(), nil, uint32(crypter.Encryptor().Overhead()))
+	ndx, err := index.Open(data.Bytes().ToByteSlice(), nil, crypter.Encryptor().Overhead)
 	if err != nil {
 		return nil, errors.Wrapf(err, "unable to open index blob")
 	}
